+++ lib/cucumber/core/report/summary.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'set'
+
 module Cucumber
   module Core
     module Report
@@ -7,7 +9,7 @@ module Cucumber
         attr_reader :test_cases, :test_steps
 
         def initialize(event_bus)
-          @previous_test_case = nil
+          @previous_test_cases = Set.new
           @test_cases = Test::Result::Summary.new
           @test_steps = Test::Result::Summary.new
           subscribe_to(event_bus)
@@ -27,8 +29,8 @@ module Cucumber
 
         def register_test_case_finished_listener(event_bus)
           event_bus.on(:test_case_finished) do |event|
-            if event.test_case != @previous_test_case
-              @previous_test_case = event.test_case
+            if !@previous_test_cases.include?(event.test_case)
+              @previous_test_cases << event.test_case
               event.result.describe_to(test_cases)
             elsif event.result.passed? || event.result.skipped?
               test_cases.flaky